Use int64_t instead of int in return value to avoid narrowing conversion

+++ include/utils/config.hpp
@@ -197,7 +197,7 @@ private:
     /// \brief loads and validates the given file \p file_path with the schema \p schema
     ///
     /// \returns the loaded json and how long the validation took in ms
-    std::tuple<json, int> load_and_validate_with_schema(const fs::path& file_path, const json& schema);
+    std::tuple<json, int64_t> load_and_validate_with_schema(const fs::path& file_path, const json& schema);
 
     ///
     /// \brief resolves inheritance tree of json interface \p intf_name, throws an exception if variables or commands
